Progress with multi metric map refactoring
