+++ doc/doxygen_project.txt.in
@@ -187,7 +187,7 @@ INCLUDE_PATH           = $MRPT_LIST_DIRECTORIES
 INCLUDE_FILE_PATTERNS  = 
 PREDEFINED             = _WIN32 MRPT_HAS_SSE2 MRPT_HAS_SSE3 MRPT_HAS_SSE4 EIGEN_PARSED_BY_DOXYGEN MRPT_HAS_PCL=1 EIGEN_MATRIXBASE_PLUGIN=<mrpt/math/eigen_plugins.h>
 EXPAND_ONLY_PREDEF     = YES
-EXPAND_AS_DEFINED      = DECLARE_COMMON_CONTAINERS_MEMBERS DECLARE_COMMON_MATRIX_MEMBERS DECLARE_MRPT_CONTAINER_TYPES DECLARE_MRPT_CONTAINER_IS_MATRIX DEFINE_MRPT_OBJECT DEFINE_MRPT_OBJECT_PRE_CUSTOM_BASE_LINKAGE DEFINE_MRPT_OBJECT_PRE_CUSTOM_BASE_LINKAGE2 DEFINE_MRPT_OBJECT_PRE_CUSTOM_BASE_NO_LINKAGE DEFINE_MRPT_OBJECT_PRE DEFINE_SERIALIZABLE DEFINE_SERIALIZABLE_PRE DEFINE_SERIALIZABLE_PRE_CUSTOM_LINKAGE DEFINE_MRPT_OBJECT_PRE_CUSTOM_LINKAGE2 DEFINE_SERIALIZABLE_PRE_CUSTOM_BASE_LINKAGE DEFINE_SERIALIZABLE_PRE_CUSTOM_BASE DEFINE_VIRTUAL_MRPT_OBJECT DEFINE_VIRTUAL_SERIALIZABLE DEFINE_MRPT_OBJECT_CUSTOM_LINKAGE  DEFINE_SERIALIZABLE_CUSTOM_LINKAGE MRPT_DECLARE_DEPRECATED_FUNCTION MRPT_DEPRECATED_PRE MRPT_DEPRECATED_POST 
+EXPAND_AS_DEFINED      = DECLARE_COMMON_CONTAINERS_MEMBERS DECLARE_COMMON_MATRIX_MEMBERS DECLARE_MRPT_CONTAINER_TYPES DECLARE_MRPT_CONTAINER_IS_MATRIX DEFINE_MRPT_OBJECT DEFINE_MRPT_OBJECT_PRE_CUSTOM_BASE_LINKAGE DEFINE_MRPT_OBJECT_PRE_CUSTOM_BASE_LINKAGE2 DEFINE_MRPT_OBJECT_PRE_CUSTOM_BASE_NO_LINKAGE DEFINE_MRPT_OBJECT_PRE DEFINE_SERIALIZABLE DEFINE_SERIALIZABLE_PRE DEFINE_SERIALIZABLE_PRE_CUSTOM_LINKAGE DEFINE_MRPT_OBJECT_PRE_CUSTOM_LINKAGE2 DEFINE_SERIALIZABLE_PRE_CUSTOM_BASE_LINKAGE DEFINE_SERIALIZABLE_PRE_CUSTOM_BASE DEFINE_VIRTUAL_MRPT_OBJECT DEFINE_VIRTUAL_SERIALIZABLE DEFINE_MRPT_OBJECT_CUSTOM_LINKAGE  DEFINE_SERIALIZABLE_CUSTOM_LINKAGE MRPT_DECLARE_DEPRECATED_FUNCTION MRPT_DEPRECATED_PRE MRPT_DEPRECATED_POST MAP_DEFINITION_START MAP_DEFINITION_END MAP_DEFINITION_REGISTER
 SKIP_FUNCTION_MACROS   = YES
 
 #---------------------------------------------------------------------------
